cahange the toast message

+++ app/src/main/java/com/example/photoday/exhibition/Exhibition.kt
@@ -39,7 +39,7 @@ object Exhibition {
             }
         } catch (e: Exception) {
             withContext(Dispatchers.Main) {
-                e.message?.let { message -> toast(activity, message.toInt()) }
+                e.message?.let { message -> toast(activity, message) }
             }
         }
     }
@@ -80,7 +80,7 @@ object Exhibition {
 
         } catch (e: Exception) {
             withContext(Dispatchers.Main) {
-                e.message?.let { message -> toast(activity, message.toInt()) }
+                e.message?.let { message -> toast(activity, message) }
             }
         }
     }
